Update js file
