+++ .eslintrc.json
@@ -21,6 +21,6 @@
   "ignorePatterns": [
     "dist/",
     "build/",
-    "src/index.js"
+    "src/"
   ]
 }
