+++ src/index.html
@@ -12,9 +12,9 @@
     <div class="wrapper">
         <div class="head-list"><h1>To do list</h1><img src="https://icon-library.com/images/reset-icon-png/reset-icon-png-2.jpg" alt=""></div>
     
-            <input type="text" id="add-task" class = "input-field" placeholder="add a new task">
+            <input type="text" id="add-task" class = "input-field" placeholder="press 'Enter' to add a new task">
             <ul class="tasks-list"></ul>
-            <button>Clear all completed</button>
+            <button class="clear-all-btn">Clear all completed</button>
     </div>
     <script src="index.js"></script>
 </body>
